scaffoldig

+++ crunch/Database/Models/PricesDaily.cs
@@ -5,13 +5,13 @@ namespace Crunch.Database.Models
 {
     public partial class PricesDaily
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateOnly Timestamp { get; set; }
         public string Symbol { get; set; }
-        public double Open { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Close { get; set; }
+        public decimal Open { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Close { get; set; }
         public long Volume { get; set; }
         public string Interval { get; set; }
         public DateTime? CreatedAt { get; set; }
